requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.28.0
aiosqlite==0.20.0
python-multipart==0.0.9
asyncpg==0.29.0
pdfminer.six
pyjwt==2.8.0
bcrypt==4.1.2
databases[postgresql,sqlite]